requirements.txt
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
email-validator=2.1.0.post1
Flask==<3.0.0>
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-JWT-Extended==<version>
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
PyJWT==2.8.0
pyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extesions=4.8.0
urllib3==2.1.0
Werkzeug ==3.0.0
uuid==1.30
WTForms==3.1.1

